+++ DDUI/src/App.jsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import './App.css';
-import "./BlogComponents";
-import PostList from './BlogComponents/PostLists';
-import CreatePost from './BlogComponents/CreatePost';
+import PostList from '@/BlogComponents/PostList';
+import CreatePost from '@/BlogComponents/CreatePost';
 
 function App() {
-    return (
-    <div className="App container">
-        <h1>Blog App</h1>
-        <CreatePost className="create-post-form" />
-        <PostList className="post-list" />
+return (
+    <div className="App">
+    <h1>DnD Blog</h1>
+    <CreatePost />
     </div>
-    );
+);
 }
 
 export default App;
